// types.ts
export interface NavItem {
  label: string;
  path: string;
}

export interface FeatureCard {
  title: string;
  description: string;
  icon: string;
}

export interface Product {
  name: string;
  scenario: string;
  image: string;
}

export interface Scenario {
  title: string;
  description: string;
  image: string;
}